requirements.txt
streamlit
numpy
Pillow
opencv-python
scikit-image
matplotlib
reportlab

